Слой Data. Интеграционные тесты репозитория [#505424]
