+++ inner/models/filter.go
@@ -1,0 +1,6 @@
+package models
+
+type Filter struct {
+	ID   int64  `db:"id"`
+	Name string `db:"name"`
+}
