+++ inner/models/role.go
@@ -1,0 +1,6 @@
+package models
+
+type Role struct {
+	ID   int64  `db:"id"`
+	Name string `db:"name"`
+}
