+++ integration_tests/repositories/.env.test
@@ -1,0 +1,6 @@
+GOOSE_DRIVER=postgres
+GOOSE_DBSTRING=postgres://postgres:password@localhost:2345/postgres
+GOOSE_MIGRATION_DIR=./migrations
+
+DB_DRIVER_NAME=postgres
+DB_DSN='host=127.0.0.1 port=2345 user=postgres password=password dbname=postgres sslmode=disable'
