+++ integration_tests/repositories/repository_test.go
@@ -12,58 +12,19 @@ import (
 	"time"
 )
 
-func TestCreateFilter(t *testing.T) {
-	a := assert.New(t)
-	db, err := database.NewDb(".env.test")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer func() {
-		if db == nil {
-			return
-		}
-		if err := db.Close(); err != nil {
-			log.Printf("error closing db: %v", err)
-		}
-	}()
-	repo := repository.NewFilterRepository(db)
-	filter := models.Filter{
-		Name: "Test Filter",
-	}
-	ID, err := repo.Add(context.Background(), filter)
-	a.NoError(err, "Failed to add filter")
-	createdFilter, err := repo.FindByID(context.Background(), ID)
-	a.NoError(err, "Failed to retrieve filter")
-	a.Equal(filter.Name, createdFilter.Name, "Filter names should match")
-}
-
-func TestFindAllFilters(t *testing.T) {
-	a := assert.New(t)
-	db, err := database.NewDb(".env.test")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer func() {
-		if db == nil {
-			return
-		}
-		if err := db.Close(); err != nil {
-			log.Printf("error closing db: %v", err)
-		}
-	}()
-	repo := repository.NewFilterRepository(db)
-	filters, err := repo.FindAll(context.Background())
-	a.NoError(err, "Failed to fetch filters")
-	a.NotEmpty(filters, "Filters should not be empty")
-}
-
 func TestCreateKey(t *testing.T) {
 	a := assert.New(t)
 	db, err := database.NewDb(".env.test")
 	if err != nil {
 		log.Fatal(err)
 	}
+	if clearErr := clearDb(db); clearErr != nil {
+		log.Printf("Error while truncating tables: %v", clearErr)
+	}
 	defer func() {
+		if p := recover(); p != nil {
+			fmt.Println(p)
+		}
 		if db == nil {
 			return
 		}
@@ -72,9 +33,13 @@ func TestCreateKey(t *testing.T) {
 		}
 	}()
 	repo := repository.NewKeyRepository(db)
+	filterID, err := NewFixtureFilter(repository.NewFilterRepository(db)).NewFilter("Empty")
+	if err != nil {
+		log.Fatal(err)
+	}
 	key := models.Key{
 		Name:     "Test Key",
-		FilterID: 1,
+		FilterID: filterID,
 	}
 	ID, err := repo.Add(context.Background(), key)
 	a.NoError(err, "Failed to add key")
@@ -118,9 +83,14 @@ func TestCreateKeyValue(t *testing.T) {
 		}
 	}()
 	repo := repository.NewKeyValueRepository(db)
+	filterID, err := NewFixtureFilter(repository.NewFilterRepository(db)).NewFilter("Empty")
+	if err != nil {
+		log.Fatal(err)
+	}
+	keyID, err := NewKeyRepository(repository.NewKeyRepository(db)).NewKey(filterID, "Kye")
 	keyValue := models.KeyValue{
 		Value: "Test Value",
-		KeyID: 1,
+		KeyID: keyID,
 	}
 	ID, err := repo.Add(context.Background(), keyValue)
 	a.NoError(err, "Failed to add key value")
